requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Async processing
celery==5.3.4
flower>=2.0.1
redis==5.0.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Performance monitoring and system utilities  
psutil>=5.9.0

# AI/LLM libraries
openai>=1.68.2
anthropic==0.8.1
litellm==1.74.8

# GitHub API
PyGithub==2.1.1
requests==2.31.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp>=3.10.0

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment management
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10
